app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
pillow==11.0.0
playwright==1.51.0
gunicorn==23.0.0